feat:整合MyBatisPlus+实现了登录注册接口

+++ zealsingerbook-framework/zealsingerbook-common/src/main/java/com/zealsinger/book/framework/common/enums/StatusEnum.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum StatusEnum {
     // 启用
-    ENABLE((byte) 0),
+    ENABLE(0),
     // 禁用
-    DISABLED((byte) 1);
+    DISABLED(1);
 
-    private final Byte value;
+    private final Integer value;
 }
